Added files via upload
When running code::blocks for the first time, it may through a fit saying that it does not have a compiler.  Don't worry, it does have a compiler, it just doesn't know it yet.  Follow the instructions in 'Compiler Problems' word document and you should never see this error again (unless you re-install code:blocks).